Move dataflowmodel to it's own public file
